+++ src/script/sigcache.cpp
@@ -26,27 +26,6 @@
 namespace
 {
 /**
- * We're hashing a nonce into the entries themselves, so we don't need extra
- * blinding in the set hash computation.
- *
- * This may exhibit platform endian dependent behavior but because these are
- * nonced hashes (random) and this state is only ever used locally it is safe.
- * All that matters is local consistency.
- */
-class SignatureCacheHasher
-{
-public:
-    template <uint8_t hash_select>
-    uint32_t operator()(const uint256 &key) const
-    {
-        static_assert(hash_select < 8, "SignatureCacheHasher only has 8 hashes available.");
-        uint32_t u;
-        std::memcpy(&u, key.begin() + 4 * hash_select, 4);
-        return u;
-    }
-};
-
-/**
  * Declare which flags absolutely do not affect VerifySignature() result.
  * We this to reduce unnecessary cache misses (such as when policy and consensus
  * flags differ on unrelated aspects).
